requirements.txt
aiosmtplib==4.0.0
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cssselect2==0.8.0
defusedxml==0.7.1
distro==1.9.0
Django==5.2.1
django-extensions==4.1
django-filebrowser==4.0.3
django-filer==3.3.2
django-filter==25.1
django-grappelli==3.0.10
django-js-asset==3.1.2
django-modelcluster==6.4
django-mptt==0.18.0
django-permissionedforms==0.1
django-polymorphic==4.1.0
django-stubs-ext==5.2.0
django-taggit==6.1.0
django-tasks==0.7.0
django-treebeard==4.7.1
djangorestframework==3.16.0
dnspython==2.7.0
draftjs_exporter==5.1.0
easy-thumbnails==2.10.1
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.12
filetype==1.2.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
laces==0.1.2
lxml==6.0.1
MarkupSafe==3.0.2
openai==1.97.0
openpyxl==3.1.5
passlib==1.7.4
pillow==11.2.1
pillow_heif==0.21.0
promise==2.3
pydantic==2.11.2
pydantic_core==2.33.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sqlmodel==0.0.24
sqlparse==0.5.3
starlette==0.46.1
svglib==1.5.1
telepath==0.3.1
text-unidecode==1.3
tinycss2==1.4.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
wagtail==7.0
wagtail-grapple==0.28.0
wagtail-headless-preview==0.8.0
watchfiles==1.0.5
webencodings==0.5.1
websockets==15.0.1
Willow==1.10.0
